fix: connection between v1 sensor edge and v3 failure analysis

+++ failure/src/main/java/ds/state/MachineState.java
@@ -40,7 +40,7 @@ public class MachineState  {
         return this.sensorsState.get(sensorId);
     }
 
-    public void updateSensorState(String sensorId, String propertyName, float value){
+    public void updateSensorState(String sensorId, String propertyName, double value){
 
     }
